requirements.txt
torch>=1.13
torchvision
pycocotools
opencv-python
matplotlib
tqdm
Pillow
